feat: Episode.java에서 usersId 자료형 통일

+++ src/main/java/com/aivle/bookapp/domain/Episode.java
@@ -28,7 +28,7 @@ public class Episode {
 
     @NotNull
     @Column(nullable = false)
-    private Long usersId;
+    private Integer usersId;
 
     @Column(nullable = false)
     @NotBlank
